# src/Web/ClientApp/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5257/"
    }
  ],
  "paths": {
...
    "/api/account/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "RegisterAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "Account"
        ],
        "operationId": "GetMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/clients/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "GetAdminClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/messages/{clientId}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "GetAdminClientThreads",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminClientThreadsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/messages/{clientId}/read-all": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "AdminMarkAllRead",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/admin/clients/{id}/reply": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "AdminReplyToClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReplyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/clients/{id}/impersonate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "StartImpersonation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminClientThreadsResponse": {
        "required": [
          "id",
          "email",
          "name",
          "threads"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "threads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreadDto"
            }
          }
        }
      },
...
      "AdminReplyBody": {
        "required": [
          "threadId",
          "subject",
          "body"
        ],
        "type": "object",
        "properties": {
          "threadId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "subject": {
            "type": [
              "null",
              "string"
            ]
          },
          "body": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "IdResponse": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ImpersonateBody": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "MeResponse": {
        "required": [
          "id",
          "email",
          "isAdmin",
          "atoConnected",
          "firstName",
          "lastName",
          "impersonating"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "isAdmin": {
            "type": "boolean"
          },
          "atoConnected": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "impersonating": {
            "type": "boolean"
          }
        }
      },
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "threadId": {
            "type": "string",
            "format": "uuid"
          },
          "inReplyTo": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "direction": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "adminRead": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RegisterBody": {
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
...
      "ThreadDto": {
        "type": "object",
        "properties": {
          "threadId": {
            "type": "string",
            "format": "uuid"
          },
          "subject": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDto"
            }
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time"
          },
          "unreadForClient": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "unreadForAdmin": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}